Remove favorites redux (#5104)
* tearing everything out

* fix favorites not persisting bug

* docs

* typo
