--- src/entities/uniswap.ts
@@ -1,3 +1,0 @@
-import { EthereumAddress, RainbowToken } from './';
-
-export type UniswapFavoriteTokenData = Record<EthereumAddress, RainbowToken>;
