--- src/references/uniswap/index.ts
@@ -1,3 +1,0 @@
-import { default as UNISWAP_TESTNET_TOKEN_LIST } from './uniswap-pairs-testnet.json';
-
-export { UNISWAP_TESTNET_TOKEN_LIST };
