--- src/references/uniswap/uniswap-pairs-testnet.json
@@ -1,1 +1,0 @@
-{}
